+++ src/main/java/me/melontini/commander/impl/expression/extensions/CustomFields.java
@@ -1,7 +1,9 @@
 package me.melontini.commander.impl.expression.extensions;
 
+import me.melontini.commander.impl.Commander;
 import me.melontini.commander.impl.expression.extensions.convert.attributes.EntityAttributesStruct;
 import me.melontini.commander.impl.expression.extensions.convert.states.StateStruct;
+import net.fabricmc.fabric.api.attachment.v1.AttachmentTarget;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
@@ -25,5 +27,7 @@ class CustomFields {
 
         ReflectiveMapStructure.addField(State.class, "properties", StateStruct::new);
         ReflectiveMapStructure.addField(LivingEntity.class, "attributes", e -> new EntityAttributesStruct(e.getAttributes()));
+
+        ReflectiveMapStructure.addField(AttachmentTarget.class, "storage", target -> target.getAttachedOrCreate(Commander.DATA_ATTACHMENT));
     }
 }